refactor: change usage of local model types on types from sdk
