+++ package.json
@@ -21,6 +21,7 @@
     ]
   },
   "dependencies": {
+    "@alexey-kuzmenko/ok-apps-sdk": "^1.2.1",
     "@emotion/react": "^11.11.1",
     "@emotion/styled": "^11.14.1",
     "@mui/icons-material": "^7.3.7",
