+++ src/components/Menu/Menu.tsx
@@ -1,13 +1,13 @@
 'use client';
 
-import styles from './Menu.module.scss';
-
 import { JSX, useEffect, useState } from 'react';
-import cn from 'classnames';
-import MenuLink from '@/types/menu-link.type';
 import Link from 'next/link';
 import { usePathname } from 'next/navigation';
+import cn from 'classnames';
+import { MenuLink } from '@alexey-kuzmenko/ok-apps-sdk';
 import { MenuIcon } from '../MenuIcon/MenuIcon';
+
+import styles from './Menu.module.scss';
 
 interface MenuProps {
     links: MenuLink[]
