--- src/models/contact.model.ts
@@ -1,15 +1,0 @@
-export enum IconType {
-    TELEGRAM = 'telegram',
-    EMAIL = 'email',
-    LINKED_IN = 'linkedIn',
-    INSTAGRAM = 'instagram'
-}
-
-export interface ContactModel {
-    _id: string
-    label: string;
-    body: string;
-    href: string;
-    iconType: IconType;
-    atl?: string;
-}
