--- src/models/skill.model.ts
@@ -1,8 +1,0 @@
-export interface SkillModel {
-    _id: string;
-    slug: string;
-    label: string;
-    iconClass: string;
-}
-
-export default SkillModel;
