--- src/types/menu-link.type.ts
@@ -1,6 +1,0 @@
-export interface MenuLink {
-    label: string;
-    href: string;
-}
-
-export default MenuLink;
